+++ frontend/src/screens/AdminScreens/UserEditScreen.js
@@ -7,7 +7,7 @@ import Loader from '../../components/Loader'
 import FormContainer from "../../components/FormContainer";
 import { getUserDetails, updateUser } from '../../actions/userActions'
 import { USER_UPDATE_RESET} from "../../constants/userConstants";
-import StickyHeader from "../../components/Header/StickyHeader";
+import HeaderDesktop from "../../components/HeaderFooter/HeaderDesktop";
 
 const UserEditScreen = ({ match, history}) => {
     const userId = match.params.id;
@@ -51,7 +51,7 @@ const UserEditScreen = ({ match, history}) => {
 
     return (
         <>
-            <StickyHeader/>
+            <HeaderDesktop/>
             <Container className='bg-white  min-vh-100'>
             <Link to='/admin/userlist' className='btn btn-light my-3'>
                 Go Back
